Make a trait that implements ModelManagerInterface
This makes it easy to implement other ModelMangers in memory
